+++ commons-http/src/main/java/com/hb0730/commons/http/support/okhttp3/OkHttp3AsyncImpl.java
@@ -125,7 +125,7 @@ public class OkHttp3AsyncImpl extends AbstractAsyncHttp {
                 if (null == httpCallback) {
                     return;
                 }
-                httpCallback.file(e);
+                httpCallback.failure(e);
             }
 
             @Override