+++ commons-http/src/test/java/com/hb0730/commons/http/support/httpclient/HttpClientAsyncImplTest.java
@@ -53,12 +53,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -77,12 +72,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -102,12 +92,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -137,12 +122,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -173,12 +153,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -195,12 +170,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -218,12 +188,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -244,12 +209,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -269,12 +229,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });
@@ -297,12 +252,7 @@ public class HttpClientAsyncImplTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
                 throw new RuntimeException("请求失败", e);
             }
         });