+++ commons-http/src/test/java/com/hb0730/commons/http/utils/HttpsTest.java
@@ -32,12 +32,7 @@ public class HttpsTest {
             }
 
             @Override
-            public void success(byte[] result) throws IOException {
-
-            }
-
-            @Override
-            public void file(Exception e) {
+            public void failure(Exception e) {
 
             }
         });