+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 ## Change
 + **[commons-cache]**
   + 更新 `Serializer#serialize` , `Serializer#deserialize` 抛出的异常信息
++ **[commons-http]**
+  + 移除 `HttpCallback#success(byte[] result)` method
 # 2.0.3
 ## Feature
 + **[commons-lang]**